OM-51712: Decreased the threshold used enable use of the quote cache.
Caching quotes is only useful if there are overlaps between cliques and
a necessary (but not sufficient) condition for that to happen is that
there are at least 2 cliques.

The condition used to be that there are at least 3 cliques for the cache
to be used and that was intentional as Shravan had measured caching was
not worth it for the old cache design and less that 3 cliques.

But for the new cache design and DXC topology I measured 9% improvement
on average when lowering the limit to at least 2 cliques. This is the
lowest limit that makes sense as different shopping lists get different
quotes even if they shop in the same market and clique.