OM-53109: Fix user validation to ignore password if the user is not a local user (rb35687)
Delivered for Christine Miller.

There is a validation check applied when editing or creating a user.  This validation check, tests whether the password is empty, but it should only do this if the user is a local user.

Tests done:
Unit tests: new unit test passed
Integration test: AD user is created correctly.