OM-53268: XL: Synchronizing jooq version across components.
As part of the group refactoring the jooq version for the group component was updated to take advantage of some newer jooq features. However, this change didn't make it to other components.

Since we use shared utilities related to databases across components, it's safer to have the jooq version be consistent. Also, the version we are using in the other components is kind of old, and updating doesn't hurt.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/35766/