OM-52955: Fix FLOW-0 used value display.
Problem
The issue here is the fact that we must make sure that only the FLOW-0 with used value of 0 gets 1. as used.
Also, we must ensure that the flow commodities are properly populated, and only when there actual flows.
And third, there's a bug fix for the case where last chunk of entity DTOs is transfered with every chunk of
topology extensions.

Solution
Only generate Flow commodities when there are actual flows.
In the Market component, set FLOW-0 used value to 1 when it is 0.
Clean up the last chunk after sending it.

Testing Done
Flows are present with the used value for the VM that has flows
The Flow congestion action is generated
history_1                       | 2019-12-06 01:46:43,826  INFO [market-api-2] [MarketListener]	: 621 entities updated

RB: 35781