OM-52695 RB:36094 Filter volumes by region using availability zones if necessary
_Background:_
In the topology, different cloud providers have different relationships between Region and Volume entities.
Azure:
Region -> CONNECTED_FROM -> Volume
AWS:
Region -> OWNS -> AvailabilityZone -> CONNECTED_FROM -> Volume

_Problem:_
Since OWNS is an "outbound" relationship and CONNECTED_FROM is an "inbound" relationship, the traversal for AWS volumes needed to specify both of them and that the intermediate step was an AvailabilityZone. This prevented Azure volumes from being traversed, because Azure has no AvailabilityZones.

_Solution:_
Both Azure Regions and AWS AvailabilityZones have CONNECTED_FROM relationships with Volumes. Instead of specifying that the traversal should start with Regions, we resolve the region filter so that if a Region is connected to AvailabilityZones, those AvailabilityZones' OIDs are used, along with the OIDs of Regions without AvailabilityZones. We use the combined group of OIDs as the starting point to query for CONNECTED_FROM Volumes.