OM-51712: Made use of quote cache configurable and disabled by default.
Measurements using the DXC (legacy) and BofA (XL) topologies showed that
enabling the quote cache has a negative performance impact for more
topologies than we initially thought. You can find details in the
comments of OM-53060 and OM-53061.

As a result we are changing quote cache mode from always-on to opt-in.

This is the `opsmgr` part of the change. There will be a corresponding
commit for the `xl` repository.