OM-53205: Accounts Greyed Out for Billing Family RB:36047 Background ============
When listing the accounts for the billing families, not all business accounts have been discovered by the user adding targets. For those accounts that have no related targets, there should be a way for the user to distinguish between those accounts.

Problem
============

The current BillingFamilyApiDTO does not support the ability to send whether the accounts related to the billing family have a related target or not. The map returned in the API response for the API endpoint `search?disable_hateoas=true&q=&types=BillingFamily` returns the following response:

```
{"accountIdToNameMap":{"323871187550":"Product Trust","001844731978":"Advanced Engineering","192821421245":"Development","631949720430":"Turbonomic SaaS"}}
```
Another part of the code, the targets page already has such a functionality in place, but the code is in a different API call, and the targets page makes follow up API calls for retreiving the related target for each account.

Solution
============

This solution is divided into three parts. This is the XL part of the solution. The goal is to have the API send an `businessUnitApiDTOList` which is an object of type `BusinessUnitApiDTO`. This allows us to send the details related to if the account has been discovered in the same response without making extra API calls:

```
{"accountDetailsList":[{"accountId":"001844731978","accountName":"Advanced Engineering","discovered":false},{"accountId":"192821421245","accountName":"Development","discovered":true},{"accountId":"631949720430","accountName":"Turbonomic SaaS","discovered":false},{"accountId":"323871187550","accountName":"Product Trust","discovered":false}]}
```