OM-53766 [RB-36188] Fix a possible NPE in StatsQueryExecuter
Background
==========
The `epoch` field was added to stat snapshots in recently to provide context regarding the stats to an API user.

Problem
=======
A helper class, `StatTimeAndEpoch` erroneously required that the Epoch parameter in its constructor be `@Nonnull` and enforced this requirements with an `Objects::requireNonNull` call. This results in a `NullPointerException` when certain sub-queries, like the Cost subquries, return stats that do not yet have an `Epoch` populated.

Solution
========
Allow the `Epoch` to be `null` in the `StatTimeAndEpoch` class.

I also added a new unit test that triggers the problem (and proves it is solved).