OM-53779, RB-36246: Search should return even when cost component is down
Intent
Search should return even when cost component is down, so that Search->BillingFamilies does not give an empty list.

Problem
Take down the cost component and visit Search->BillingFamilies on the UI. The list will be empty, even though there's billing families from the AWS Development target.

Diagnosis
Cost component is down, so GRPC throws a StatusRuntimeException which extends RuntimeException. The exception is uncaught and travels all the way down the stack.

Solution
Catch StatusRuntimeException and check if the code is Code.UNAVAILABLE. If so, log a message and use a the approriate datastructure to communicate that the data is unavailable. Ensure that null cost price is returned in business unit dto and billing familiy dtos.

Alternative Solutions
1. Return 0 cost
pros: much simplier solution since no null checking needed
cons: poor user experience. the customer cannot distinguish between true 0 cost (nothing to pay) and cost component unavailable
2. All api responses should return a component describing missing information
pros: provides the best customer experience because it has the potential to communicate to the customer what component has issues
cons: beyond the scope of work that can fit into a bug fix