OM-53782,RB-36279-Fix the issue that the discount applied to a billing family is only applied to master account
Intent
======
Add new discounts for children accounts as we add a discount for top accounts

Problem
=======
When you create a price adjustment for a billing family , the discount is only applied to the master account.

Diagnosis
=========
The logic for adding the discount to child account of a master account is missing.

Solution
========
Based on discussion with Evan, the temporary solution for this problem is that in the API we get the children accounts when creating a discount and create discount for them.