OM-53996: Improve NCM support for containers.
iThe NCM for Containers treated the ContainerPods as yet another layer on top of:
PM -> VM
The issue with that approach is that physical hosts are a simulated resource in the Cloud, and therefore can't be relied upon for reliable distance calculation.
The Communication Matrix requires reliable capacity/latency calculation based on the underlay (actual physical) network.
In container world, especially in the cloud, it makes more sense to use Nodes/Clusters as a basis for such a calculation.
So, the change is required to treat distances as follows:
Was
Blade - Flow 0
Single hop (TOR Switch) - Flow 1
Same Datacenter - Flow 2
Cross Datacenter - Flow 3

Now
Same Kubernetes Node - Flow 0
Same Kubernetes Cluster - Flow 1
Different Kubernetes Clusters - Flow 2 (provided they are in the same region)
Different Datacenters/Regions - Flow 3

What will remain to be tested/modified as needed is support for cross-region traffic.
That problem is non-trivial, as while different regions have the same capacity class, they will most likely have different latency or in fact, different capacity.
That should be left, in my opinion, to a later stage.

RB: 36310