OM-51712: Refactored a QuoteCache test to reduce its complexity.
The `testPutGet_emptyCache_sequence_positive` parametrised test was
refactored to reduce its cyclomatic complexity by making it more
abstract. ie. The test asserts that adding 3 cache entries works as
expected and the old code was specialised for that number. The new code
still tests the same scenarios, but is agnostic to the number of
entries, has reduced cyclomatic complexity (11 instead of 17) and is
slightly sorter (by 6 lines).

This was triggered by a comment by ReviewBot, although it should be
noted that the code in question was test code, not production code.