OM-53920,RB-36315-Fix the issue that the IP cost for IDLE vms are not being shown properly.
Intent
======
Consider that the VMs don't have free IPs when they are idle in the cost calculation.

Problem
=======
The AWS VMs get a free IP when they are running. However, AWS will charge for them when the VM is idle. However, we do not show that when scoping to idle vms.

Diagnosis
=========
The logic for that is missing.

Solution
========
Add a logic in IP cost calculation which sets the free ips to zero when the VM is not powered on.