OM-53266,RB-36357: Hide unmonitored business accounts in search and list of accounts
Intent
Hide unmonitored business accounts in search and list of accounts so that counts and visibility is consistent with classic and top accounts.

Problem
Looking at a Billing Family, the list of accounts said the aws engineering account had 4 accounts, but the Cloud Service Provider widget only had one.

Diagnosis
List of Accounts was returning all accounts in the billing family, even the ones that were not monitored. In CSP widget we purposely hid the accounts without a target in this JIRA: OM46346 by this review https://rbcommons.com/s/VMTurbo/r/31409/ so that the count would match what's in the Top Accounts widget. The Top Accounts widget only shows the monitored accounts.

Solution
Do not count business accounts that do not have an associated target id.
Do not return business accounts that do not have an assoicated target id.

Unmonitored business accounts should only be visible when you expand the BillingFamily in Search->BillingFamily or when you expect the related target that found it in Settings->Target Configuration

Alternative Solutions
Fix the UI instead of API
Con: Multiple locations will need to be fixed, instead of just a couple in API
Con: Counts cannot be fixed would be need to be replaced by a list, filter, then size of filtered list

Unrelated fixes
one spotbug fix
a bunch of checkstyle fixes