OM-53756: XL TP/Group: Dedupe scanned groups/setting policies, and improve handling of future duplicate groups.
There are three parts to the fix:
1) In the StitchingGroup class, don't throw a runtime exception if a single group is discovered multiple times by the same target. The behaviour for now is: take the first group. If the duplicates contain no new members, print a warning. If the duplicates contain new members, print an error.
2) In the GroupInterpretationContext, don't throw a runtime exception if the probe uploads two groups with the same id. If the groups are NOT the same, choose the first and print an error.
3) In the DiscoveredGroupScanner, avoid creating scanned groups or setting policies with the same name.

There is also an improvement to the DiscoveredGroupUploader. We move all the maps of target id to groups/policies/setting policies to a helper class - **TargetDiscoveredData** - and split the scanned groups and setting policies from the discovered ones. Every target discovery creates a new partially immutable **TargetDiscoveredData** and associates it with the target. Every broadcast, the DiscoveredGroupScanner overrides the scanned groups/policies for relevant targets with the newly scanned ones. Diagnostic collection only takes the discovered groups/policies/setting policies, because the scanned ones are implicitly contained in the entities.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/36300/