OM-53646,RB-36465-Don't set the entity type filter as empty list if there is no entity
Problem
=======
The on-demand cost for DB after action is the same as on-demand cost before action

Diagnosis
========
When asking for the cost of the DB, we have a logic which creates the entity type filter and if there is a entity type in the request it will add it to the filter list. This results in an entity type filter which is an empty list. Therefore, we are essentially say only return the costs for entities which has a type in the list of empty list which returns nothing. we have a logic which returns current cost when empty is returned by projected entity cost store.

Solution
========
We should not set entity type filter if there is no entity type in the request.