OM-54302 (and 54313); [RB-36536]
Problem
=======
In XL when showing aggregated stats for a group, the capacity field contains the capacity of a random member of the group. Instead, it should contain the aggregated capacities of all the group members.

Solution
========
Aggregate the capacities during stat aggregation--instead of replacing the old value with the new one, add them together.