OM-54014: XL Market: Restrict number of concurrent analyses.
Introducing a "TopologyProcessingGate" which controls how many analyses are allowed to run concurrently. For 7.17.9 the gate implementation we will use is a simple one that limits concurrent plans to 1 (configurable). For 7.21 we can add "smarter" logic which would try to guess how many entities the market component can be analyzing at a time (maybe based on heap size), and only allow that many analyses. The "TopologyProcessingGate" is a shared utility - we can use the interface in other components that take in and process topologies too.

Also, I noticed that we don't print the ID of the topology used in a plan. Addressing that with a small fix to return the topology ID in StartAnalysisResponse.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/36327/