OM-54579 [RB-36673] Filter entities from plan summary stats in a consistent way
Problem
=======
When we filter suspended entities from the plan results, we filtered them only from the *counts* but not from the aggregated commodity stats. When entities are suspended, the results show the count decreasing but the stats remaining the same.

Solution
========
Move the spot where we do the filtering so that it impacts both the counts and the stats aggregation.