OM-54030 [RB-36689] Support groupBy and peak values for plan entity stats.
Background
==========
The API exposes an optional `groupBy` parameter. In Classic, if this parameter was supplied then results would be aggregated into groups based on the value of the parameter. If `groupBy` is omitted, the stats for all commodities of the same type would be aggregated together.

Problem
=======
In XL, we did not have any support in Repository for `groupBy`. The impact of this is that requests for per-entity plan data would disregard the parameter. Worse, the default behavior in XL does not match the default behavior in Classic, to aggregate all commodities of the same type together. Instead, in XL we were implicitly grouping by `key` by default. This causes the API callers (including the UI) to get results aggregated in unexpected ways.

Solution
========
Introduce an initial implementation of `groupBy` in Repository component for plan entity stats. Also, change the default behavior to match Classic.

Note: This review also fixes issue #54664 as a side effect. Consider it an added bonus!