OM-53806- Make the migration script V_1_26 idempotent
Problem
=======
For some appliances the cost pod goes to CrashLoopBackOff

Diagnosis
=========
The migration v1_26 for cost component can take a while if the size of cost tables are big. There is always a possibility that a pod restarts for different reasons. If the pod restarts in the middle of migration that will result in the half-executed migration which prevents the execution as some of its commands has been previously executed and results in error if we try to run them again.

Solution
========
I made changes to that migration script to be idempotent. This means if the script gets executed half-way through, it can start again and does not result in broken migration. This will ensure that never end-up with broken pod. However, this means that the checksum changes for that migration script which can break the already executed migrations.  To fix this, I will use the solution used in https://rbcommons.com/s/VMTurbo/r/34180 which uses flyback callback to fix the checksum.