OM-54277,RB-36891-Ignore the expired RI when scoping to a topology
Intent
======
Change the GetReservedInstanceBoughtByTopology GRPC Service so it does not return the expired RIs

Problem
=======
The market makes recommendation to utilize expired RI.

Diagnosis
=========
When the market asks for RIs in an existing toplogy, we don't filter the expired RIs

Solution
========
Filter the expired RI when calling that method. It is not possible to do that in the sql query since this information is kept as a BLOB. Therefore, the filter is performed after retrieving the RIs.