OM-54692 [RB-36918] Set the commodity relation (bought or sold) on plan entity stats. Also ensure that max value is not less than average or min values.
Problem
=======
For plan stats, we were not properly setting the commodity relation field to either "bought" or "sold". This means that in cases where the bought and sold commodities have the same name (as in the case of storage), the UI did not know which commodity stats to display.

Additionally, we did not have safeguards to ensure that the "max" stat was equal to or greater than the other (average and min) stats. In some cases, the max is set to whatever "peak" is set to on the entity--and this can be unset (zero) in some cases.

Solution
========
Set the commodity relation (bought or sold) on plan entity stats. Also ensure that max value is not less than average or min values.

I did some minor cleanup (to replace references to the "UI" with references to the "API") and updated some comments.