OM-52785 [RB-36944] Do not include unplaced entities in plan result counts or widgets
Background
==========
When running a plan it is possible for an entity (usually a VM) to become unplaced. This means that it could not find providers to buy one or more of its commoditiesBought from. We keep a separate count of unplaced VMs, and display a gray banner in the UI with the count when any are present in the plan results.

Problem
=======
Unplaced VMs (and in theory, other unplaced entities) are currently being included in all the plan results counts, densities, charts and stats... just as if they'd been placed. This is confusing to the user, who wants to see how many entities were actually placed successfully in the counts, densities, charts and stats.

We have the further complication of partially placed entities, which were placed on one or more providers but were not placed for other providers (see Phase 2 below).

Solution
========
We reached a consensus of having 2 phases:

Phase (1) 7.21.0 (this review)

    Fix the summary table. If a VM is not placed it’s NOT going to be counted as part of “after” or be calculated in host&storage density
    Unplaced VMs won’t appear in “VM Optimized Improvements” or “VM Comparison”

Phase (2) 7.21.1 https://vmturbo.atlassian.net/browse/OM_54889

    Partial unplaced VMs won’t affect utilization of hosts&storage
    Partial unplaced VMs won’t appear in “VMs per Host” and “VMs per Storage” in “Host Comparison” and “Storage Comparison” widgets.

This review implements Phase 1:

* Changes in the History component suppress unplaced entities from the summary statistics and counts
* Changes in the Repository exclude unplaced entities from the detail widgets (like Optimized Improvements)