OM-55030 [RB-37036] For plan summary results, aggregate StorageAmount commodity only for Storage devices
Background
==========
Plan summary stats (both source and projected) are calculated by the History service on ingestion (i.e. when receiving the topology from the Market) and the aggregated totals are stored in the database. For commodities, the commodities sold are aggregated on a per commodity type basis. Market-wide commodity aggregations can then be requested by name by API users. For the purposes of these plan summary stats, we do not support the usual optional filters for stats requests like groupBy or filtering by entity type.

Problem
=======
We noticed that there is some double (and even triple) counting going on for the StorageAmount commodity. The reason for this is that the same exact CommodityType is bought and sold at several layers of the Topology. For example, a Storage Device, a Disk Array and a Storage Controller all sell StorageAmount. Since our aggregation logic for commodities aggregates only by CommodityType, the amounts would be aggregated across all entities that sell this commodity, regardless of their entity type. This leads to much higher storage totals than expected.

Solution
========
For 7.21.0, we introduce a low-risk fix. For the StorageAmount commodity specifically, we will filter out entities from the aggregation that are not Storage Devices.

Long-term, it would be better to store the aggregations on an entityType/commodityType grid, rather than simply storing them by commodityType. This would give us more flexibility with the returned data. There is a linked issue created for this.