OM-54277-Ignore the expired RIs in all queries made related to RIs bought
Intent
======
Ignore the expired RIs in all operations related to RI

Problem
=======
We consider expired RIs costs in the reserved computed cost in CCC.

Diagnosis
=========
We do not filter based on expiry time of RIs when we are  querying for the aggregated RI cost.

Solution
========
The start time of RI of an RI is kept as in a serialized protobuf object in database. The expiry time is calculated based on start time, the term of the reserved instance specification which is kept in another table. Therefore, querying based on it expiry time is not possible right now. Therefore, I added new column for reserved instance start and expiry time in the database which can be used in querying so we can filter the expired RIs. Since existing data does not have those columns populated, I created a migration script which cleans up ``reserved_instance_bought`` and ``entity_to_reserved_instance_mapping`` tables. These tables will be populated with correct data in the next broadcast with columns populated. The value of RI start and expiry date is populated in the cost component.