OM-55076: XL TP: Improving reservation robustness and handling POLICY and non-POLICY constraints.
Two issues:
1) When a reservation has a non-policy constraint, we look through it's constraints to figure out the providers to use to create a Bind policy. However, if these constraints also include a POLICY constraint there is a block of code that ends up throwing an illegal exception.
2) If a reservation plan fails to broadcast for whatever reason, the reservation manager doesn't know about it, and has no way of marking the reservation invalid. This means the reservation stays "IN PROGRESS" and continues to block subsequent broadcasts.

Fix
===
For 1), we just need to NOT throw the exception on POLICY constraints. We handle POLICY constraints separately.
For 2), we mark plans as failed when the topology broadcast for the plan fails. I'm also making the ReservationManager listen for plan failure notifications, so it can handle any kind of plan failure properly.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37061