OM-54904 [RB-37135] Delete obsolete plans when migrating to 7.21
Summary
=======
Per Mary Nachimov, we are now planning to not support migration of plans between 7.17 and 7.21. Too much has changed, and even if we could migrate all the data successfully, the end result of displaying data from legacy plans would not make sense in the context of Plan Overhaul.

This migration will purge all existing plans when upgrading to 7.21 which includes Plan Overhaul, multi-tenancy and other breaking changes/improvements.