OM-55288: XL TP: Speed up EnvironmentTypeInjector for applications.
Speed up the calculation by introducing a cache for the environment types below (and including) an entity in the supply chain. I (over)-optimized the cache by using a RoaringBitmap to represent the providers that have the environment type. This is slightly complicated, but fairly isolated, and results in large savings - the size of the cache was 0.5MB when populated with 200k entities. The straightforward approach (using Set<Long> for the providers, and a Map) took 16MB!

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37163/