OM-51712: Added a default value in an @Value annotation for a setting.
The design of our current settings-loading sub-system was that such
annotations shouldn't be included and instead defaults be specified in
only one place. Unfortunately, operator images aren't guaranteed to
match component images at the moment and until OM-53975 is resolved we
have to guarantee somehow that components won't crash when a default is
missing from the configMap.

In order for a component to not crash at startup at least one of the
following must hold:
 - The operator version running on the instance has been compiled with
   an up-to-date versions of
   `kubernetes/operator/helm-charts/xl/templates/configmap.yaml`.
 - `/opt/turbonomic/kubernetes/operator/deploy/crds/charts_v1alpha1_xl_cr.yaml`
   on the instance has been modified to include an override for the
   setting.
 - The `@Value` annotation loading the setting's value in Java includes
   a default.

Note that in the latter case, the setting's existence and value won't be
properly printed at component startup, but it's the only one of the
conditions that we can control.

Also loaded the 2019-12-17 RBC topology (converted to 7.21.x) and
measured a 30% reduced analysis time with quote cache enabled compared
to disabled, proving that there are some topologies that benefit from
cache even if they aren't many.