OM-55508, RB-37412: Severity calculations should also consider severity breakdown
Intent
======
Severity calculations consider entity severity and severity breakdown so that the customer can see there's an potential issue with the application, in list contexts that are due to the underlying host, virtual machine, storage, etc.

Solution
========
Modify severity populator to consider entity level severity and severity breakdown when getting the severity and when calculating the severity for a group.

pro: only one file to change
con: UI cannot tell if the severity is coming from entity level or severity breakdown. this is acceptable because the UI doesn't have a usecase for the difference

Alternative Solution
====================
1. Send both severity and severity breakdown to the UI and let it figure out whihc
    1. con: lots of places in the UI to change