OM-53111: XL AO: Setting prior and new action count for monthly action stat records.
The monthly action stat columns don't have prior or new counts, because we assumed that we wouldn't need to roll them up. However, we use the prior and new counts to populate the "total" action count for the month. Having the values be 0 means the "total" actions always appear as 0.

Fix
===
Add the missing columns to the database, and populate them like we do with other (hourly, daily) tables.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37468/