OM-55812: Made Market constructor package-private again.
Market objects can only exist inside an Economy and should only be
created and destroyed through the API of the latter class. In order to
enforce that the Market constructor must be package-private.

Yet it was made public to use Mockito in a test. Mockito is a mocking
framework and analysis library classes are designed to be testable
without the need to mock them. Thus I'm restoring the old accessibility
and creating the market object in the appropriate way.

Also cleaned up imports to comply with Turbonomic coding standards.