OM-55812: Refactored QuoteSummerTest to not use Mockito.
Mockito is a mocking framework. It's useful for testing classes and
methods that reference classes (or methods) that can't be reasonably
created (or called) within the constraints of unit tests. e.g. they may
require interacting with other processes, rely on global state or just
not be designed in a testable way.

On the other hand, tests written using Mockito require and encode
precise knowledge of how the method-under-test is implemented (to mock
all the methods it calls internally and with the correct arguments).
This is in contract to the principle of writing tests for the interface
and not the implementation. (And also means that if the implementation
changes, the test will also have to change)

Tests using mocking also require and encode the contracts of all the
methods the method-under-test calls, not just contracts of the method-
under-test.

Last but not least, it's easy to mock classes in an invalid way that
violates their contracts. e.g. in the case of QuoteSummerTest, 3 traders
with the same economy index were created although a real Economy object
would never allow that.

The analysis library has been designed from the ground up to be testable
without the need to mock its classes. So to avoid the above pitfalls the
mocking code was removed. It should be noted that the non-mocked code is
shorter.

Also:
- Made fields private
- Moved all initialisations into `@Before` method.
- Added some whitespace to comply with Turbonomic coding standards.