OM-55665 [RB-37571] Update ArangoDB to 3.6.1
Note: As part of this upgrade, we are migrating the Arango storage engine from 'mmfiles' format to 'rocksdb' format. The reason is that mmfiles is now deprecated. I wrote some bash scripts (automatically run by the docker container entrypoint, as needed) to automaticallly migrate the user data to the new format. Therefore, plan data should persist through this upgrade.