OM-55813: XL: Add timestamp to component instance IDs for consul registration.
When a component starts up it registers with Consul using the k8s pod name as the instance ID. When the JVM shuts down it deregisters the instance ID. However, when pods restart the "new" pod uses the same name as the old one, and it may be possible that there are overlaps in the shutdown of one JVM and the startup of another. This may cause the old pod's shutdown to clear the registration of the new pod.

Fix
===
The fix is to add the current time to the instance ID. This will ensure that the post-restart pod and pre-restart pod register with different instance IDs in consul. The shutdown hook will work as before, so we won't be increasing the risk of polluting consul with old instance IDs.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37578/