OM-55673: on-prem resize down limited by percentile usage instead of max used
Intent
======

On-prem resize down limits by percentile usage insetad of max used so that optimize on-prem will produce similar vmem capacity recommendations to migrate to cloud.

Problem
=======
PwC with Nakul's help noticed that on-prem we resized VMem from 16GB to 15GB, even though max rate of resize was set. Migrate to cloud picked a 4GB machine instead. PwC and Nakul were unsettled by this large difference.

Diagnosis
=========
1. Diagnostics customer.repo.topology file showed that maxUsed was 15GB.
2. `Max used*0.9` set to maxQuantity on CommoditySoldTO
3. Resizer limits resize down to maxQuantity

Solution
========
Replace maxUsed with percentile utilization in market 2 Resizer
pros: resize behavior of on-prem will match migrate to cloud
cons: tests are difficult to setup

Limitations
===========
Before and after fix, on-prem mgirated 6.5GB to t3 nano which is 0.5GB. This is incorrect because in both cases, it is resizing below heap capacity. This issue can be filed as a separate bug.

Alternative Solution
====================
Explain to customer that on-prem resize down.
1. pro: no work - no testing need
2. con: poor customer experience
3. con: on-prem does not match with migrate to cloud
4. con: percentile usage is dead code for on-prem

Unit Test
=========
1. When historical and max quantity or both set, resize should go below max quantity, but not below historical quantity.