OM-56030: XL API: Don't throw exception when group has no entities in supply chain.
We try to create a group of "All On-Prem Hosts" (because that's what the UI looks for). However, there are no hosts in the system. There's a place in the code where we expect to find entities, and we should handle the empty case (because it's valid).

Fix
===
Instead of throwing exception, treat missing types in the supply chain as "empty".

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37824/