OM-53517: XL: Upgrade gRPC, netty, and protobuf to latest versions.
Issue
=====
Occasional resource leaks at the Netty/gRPC level: LEAK: ByteBuf.release() was not called before it's garbage-collected.

Background
==========
It seems like this is something that can happen intermittently when Netty's built-in leak detection detects an issue.

Fix
===
It's hard to diagnose because this seems like it's under the application layer. There may be something in how we're using gRPC that causes this, or it could be a gRPC (or even Netty) issue. The first step is to upgrade gRPC (and, implicitly, Netty) to the newest version. There have been some memory leak related fixes between the version we have and the latest one, and it's possible that this will go away.

Note: The mediation.gcp dependency has its own version (transitively) of gRPC dependencies. I'm excluding them in the imports so that we don't end up with different versions in the JAR.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37835/