OM-56033: XL Tests: Use DbCleanupRule and DbConfigurationRule instead of TestSQLDatabaseConfig for db-related unit tests.
The idea was that **DbConfigurationRule** would be a **ClassRule** that runs once per class to set up the database and tear it down, and **DbCleanupRule** would run before every test to remove all data in the database. This means that you save (num of tests in class - 1) migrations. There are additional optimizations we can try to make in the future (e.g. reorganize into suites, to do one migration per component), but this was mainly taking an existing solution and applying it broadly to all components. There are also improvements we can make to other tests - e.g. communication.matrix takes 1.5 minutes because there are a bunch of tests that sleep for multiple seconds. Reserved instance allocator tests are also suspiciously long (takes as long as the topology-processor).

Note - this change means that tests within a class are not safe to run concurrently. The database created by the DbConfigurationRule is shared for the test class, so since we will be cleaning up a shared database we need the class tests to run sequentially. If we want the tests to run concurrently we should use the surefire plugin instead of "-T*".

Reviewed At: https://rbcommons.com/s/VMTurbo/r/37670/