OM-56521: Refactored ReplayActionsTest to avoid accessing private fields
`ReplayActionsTest` was using reflection to bypass protection and
manipulate private fields of `Topology` in a way that violated the
invariants of the latter. The end result was that, although the tests
passed for a particular implementation of `ReplayActions`, they wouldn't
pass for any alternative implementation.

Now the test should correctly allow the implementation of `ReplayActions`
to change as long as the interface and contracts remain the same.

Also:

- Some looping constructs in the tests were just there to return the
  first element of a linked set. These were replaced by simpler
  statements. More refactoring is needed to get the desired shopping
  list without iterating over all of them (and that will have to happen
  at a later time) but at least now the iteration is internal to Java's
  `toArray` and not copied and pasted repeatedly.
- Simplified some assertions at IDE's suggestion.
- Renamed some variables.